requirements.txt
async-timeout==4.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
exif==1.6.0
flickrapi==2.4.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.1
parameterized==0.9.0
Pillow==9.5.0
pluggy==1.0.0
plum-py==0.8.6
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
PyExifTool==0.5.5
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pytest==7.3.1
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
